Claude animations for boon
I saw this in an aricle - particle.html
create an app in one html page that can be used in Artifacts. make an interactive
3D particle cloud, with a max of 100 particles. use Three.js for the simulation

and followups

please add easy to change variables at the top of the page for controlling particle color size and speed  - thanks

can we add a variable for view zoom



and I twekaed the config values
        // Easy to change variables
        //const PARTICLE_COUNT = 100;
        //const PARTICLE_SIZE = 0.1;
        //const PARTICLE_COLOR = '#ffffff'; // Use null for random colors
        //const ROTATION_SPEED = 0.002;
        //const MOUSE_SENSITIVITY = 0.5;
 
        const PARTICLE_COUNT = 1000;
        const PARTICLE_SIZE = 0.04;
        const PARTICLE_COLOR = null //'#39FF14' //'#ffffff'; // Use null for random colors
        const ROTATION_SPEED = 0.05;
        const MOUSE_SENSITIVITY = 0.5;        
        const CAMERA_ZOOM = 10; // New variable for camera zoom (higher values zoom out)





and so I tried to create one for boons amber portal project animation
clustering2.html
create an app in one html page that can be used in Artifacts. make an interactive 3D chart of a machine learning  progression over time of  the clustering of a million rows of 40 columns of random data . use Three.js for the simulation.  Add variables at the top of the page to control number of rows, columns, speed of visualization , and view zoom level please

follow up
can we add a camera rotation with a variable for the speed of that as well please